+++ src/qq_lib/jobs/cli.py
@@ -33,13 +33,19 @@ logger = get_logger(__name__)
     help="Username whose jobs should be displayed. Defaults to your own username.",
 )
 @click.option(
+    "-e",
+    "--extra",
+    is_flag=True,
+    help="Show additional information about the jobs.",
+)
+@click.option(
     "-a",
     "--all",
     is_flag=True,
     help="Include both unfinished and finished jobs in the summary.",
 )
 @click.option("--yaml", is_flag=True, help="Output job metadata in YAML format.")
-def jobs(user: str, all: bool, yaml: bool) -> NoReturn:
+def jobs(user: str, extra: bool, all: bool, yaml: bool) -> NoReturn:
     try:
         BatchSystem = QQBatchMeta.fromEnvVarOrGuess()
         if not user:
@@ -56,7 +62,7 @@ def jobs(user: str, all: bool, yaml: bool) -> NoReturn:
             sys.exit(0)
 
         BatchSystem.sortJobs(jobs)
-        presenter = QQJobsPresenter(jobs)
+        presenter = QQJobsPresenter(jobs, extra)
         if yaml:
             presenter.dumpYaml()
         else: